+++ src/lib/playground/proximity/stores.ts
